# backend/src/swagger.yaml
openapi: 3.0.0
info:
  title: My Food Ordering API
  description: An API for managing food orders
  version: 1.0.0
servers:
  - url: http://localhost:5000/api
paths:
  /foods:
    get:
      summary: Get all foods
      responses:
        '200':
          description: A list of foods
    post:
      summary: Create a new food
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                ingredients:
                  type: string
                price:
                  type: number
                imageUrl:
                  type: string
                meat:
                  type: boolean
                tag:
                  type: string
      responses:
        '200':
          description: The created food
  /foods/{id}:
    get:
      summary: Get a food by ID
      parameters:
        - name: id
          in: path
          required: true
          description: The ID of the food to retrieve
          schema:
            type: string
      responses:
        '200':
          description: The requested food
  /users/login:
    post:
      summary: Log in a user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                password:
                  type: string
      responses:
        '200':
          description: Logged in successfully
  /users/register:
    post:
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                email:
                  type: string
                password:
                  type: string
                address:
                  type: string
      responses:
        '200':
          description: Registered successfully
  /orders/create:
    post:
      summary: Create a new order
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                items:
                  type: array
                  items:
                    type: object
                    properties:
                      food:
                        type: object
                      price:
                        type: number
                      quantity:
                        type: integer
                name:
                  type: string
                address:
                  type: string
                paymentId:
                  type: string
      responses:
        '200':
          description: The created order
  /orders/all:
    get:
      summary: Get all orders for the authenticated user
      responses:
        '200':
          description: A list of orders
  /orders/{id}:
    get:
      summary: Get an order by ID
      parameters:
        - name: id
          in: path
          required: true
          description: The ID of the order to retrieve
          schema:
            type: string
      responses:
        '200':
          description: The requested order